requirements.txt
beautifulsoup4==4.12.3
ipython==8.12.3
langchain==0.3.14
langchain_anthropic==0.3.1
langchain_community==0.3.14
langchain_core==0.3.29
langchain_openai==0.3.0
langchain_teddynote==0.3.37
langchain_text_splitters==0.3.5
Markdown==3.7
matplotlib==3.10.0
pandas==2.2.3
Pillow==11.1.0
pydantic==2.10.5
pymupdf==1.25.1
python-dotenv==1.0.1
react_agent==0.0.1
Requests==2.32.3
retriever==3.1.0
streamlit==1.41.1
typing_extensions==4.12.2
